backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
aiofiles==23.2.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0
PyPDF2>=3.0.0
python-docx>=0.8.11
chardet>=5.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.24.1
requests==2.31.0